+++ src/skin/mod.rs
@@ -1,5 +1,7 @@
 mod app_skin;
 mod cli_mad_skin;
+pub mod colors;
+mod ext_colors;
 mod help_mad_skin;
 mod panel_skin;
 mod purpose_mad_skin;
@@ -10,6 +12,7 @@ mod status_mad_skin;
 pub use {
     app_skin::AppSkin,
     cli_mad_skin::*,
+    ext_colors::ExtColorMap,
     help_mad_skin::*,
     panel_skin::PanelSkin,
     purpose_mad_skin::*,